Reverse dependencies between api and transport
