+++ crates/bcr-ebill-api/Cargo.toml
@@ -27,7 +27,6 @@ futures.workspace = true
 infer = { version = "0.19.0", default-features = false }
 bcr-ebill-core.workspace = true
 bcr-ebill-persistence.workspace = true
-bcr-ebill-transport.workspace = true
 tokio.workspace = true
 tokio_with_wasm.workspace = true
 secp256k1.workspace = true
