+++ crates/bcr-ebill-api/src/service/bill_service/error.rs
@@ -1,4 +1,4 @@
-use crate::{blockchain, external, persistence, util};
+use crate::{blockchain, external, persistence, service::notification_service, util};
 use thiserror::Error;
 
 /// Generic error type
@@ -33,7 +33,7 @@ pub enum Error {
     Cryptography(#[from] util::crypto::Error),
 
     #[error("Notification error: {0}")]
-    Notification(#[from] bcr_ebill_transport::Error),
+    Notification(#[from] notification_service::Error),
 
     #[error("io error {0}")]
     Io(#[from] std::io::Error),
