+++ crates/bcr-ebill-api/src/service/bill_service/service.rs
@@ -25,6 +25,7 @@ use crate::persistence::company::{CompanyChainStoreApi, CompanyStoreApi};
 use crate::persistence::contact::ContactStoreApi;
 use crate::persistence::file_upload::FileUploadStoreApi;
 use crate::persistence::identity::{IdentityChainStoreApi, IdentityStoreApi};
+use crate::service::notification_service::NotificationServiceApi;
 use crate::util::BcrKeys;
 use crate::util::file::UploadFileType;
 use crate::{external, util};
@@ -47,7 +48,6 @@ use bcr_ebill_core::notification::ActionType;
 use bcr_ebill_core::util::currency;
 use bcr_ebill_core::{File, ServiceTraitBounds, Validate, ValidationError};
 use bcr_ebill_persistence::mint::MintStoreApi;
-use bcr_ebill_transport::NotificationServiceApi;
 use log::{debug, error, info, warn};
 use std::collections::{HashMap, HashSet};
 use std::sync::Arc;
