+++ crates/bcr-ebill-transport/src/handler/company_invite_handler.rs
@@ -1,6 +1,10 @@
 use std::{cmp::Reverse, collections::HashMap, str::FromStr, sync::Arc};
 
 use async_trait::async_trait;
+use bcr_ebill_api::service::notification_service::{
+    event::{ChainInvite, ChainKeys},
+    transport::NotificationJsonTransportApi,
+};
 use bcr_ebill_core::{
     NodeId, ServiceTraitBounds,
     blockchain::{BlockchainType, company::CompanyBlock},
@@ -11,10 +15,11 @@ use bcr_ebill_persistence::{NostrChainEventStoreApi, nostr::NostrChainEvent};
 use log::{debug, error, warn};
 
 use crate::{
-    Event, EventEnvelope, EventType, NotificationJsonTransportApi, Result,
-    event::blockchain_event::{ChainInvite, ChainKeys},
+    EventType,
     handler::public_chain_helpers::{BlockData, EventContainer},
 };
+use bcr_ebill_api::service::notification_service::event::Event;
+use bcr_ebill_api::service::notification_service::{Result, event::EventEnvelope};
 
 use super::{
     CompanyChainEventProcessorApi, NotificationHandlerApi,
@@ -176,7 +181,6 @@ impl CompanyInviteEventHandler {
 #[cfg(test)]
 mod tests {
     use crate::{
-        event::blockchain_event::CompanyBlockEvent,
         handler::{
             MockCompanyChainEventProcessorApi,
             public_chain_helpers::collect_event_chains,
@@ -185,10 +189,12 @@ mod tests {
                 private_key_test,
             },
         },
-        transport::{MockNotificationJsonTransportApi, create_public_chain_event},
+        test_utils::MockNotificationJsonTransport,
+        transport::create_public_chain_event,
     };
 
     use super::*;
+    use bcr_ebill_api::service::notification_service::event::CompanyBlockEvent;
     use bcr_ebill_core::{
         blockchain::{
             Blockchain,
@@ -410,12 +416,12 @@ mod tests {
     }
 
     fn get_mocks() -> (
-        MockNotificationJsonTransportApi,
+        MockNotificationJsonTransport,
         MockCompanyChainEventProcessorApi,
         MockNostrChainEventStore,
     ) {
         (
-            MockNotificationJsonTransportApi::new(),
+            MockNotificationJsonTransport::new(),
             MockCompanyChainEventProcessorApi::new(),
             MockNostrChainEventStore::new(),
         )
